+++ R/crssi_change_nf_start_date.R
@@ -25,24 +25,46 @@ seperateHeaderFromData <- function(x) {
   return(list("data" = theData, "atts" = atts))
 }
 
-#' changeStartDate
+#' Change the start dates of CRSS input trace files
 #' 
-#' changes the start date in all files in each trace folder
-#' 
-#' Used to change the start date of all natural flow input files. Will change them for all 
-#' files in a folder.
+#' `crssi_change_nf_start_date()` changes the start date in all files in each 
+#' trace folder. Used to change the start date of all natural flow input files. 
+#' This Will change the start date for all files in a folder (flow, salinity, 
+#' other).
 #' 
 #' @param nTrace The number of traces to process.
-#' @param folder A string with either a relative or absolute path that contains the trace 
-#' folders, e.g., 'C:/CRSS/dmi/DNFSinput'
-#' @param startDate string of new starting date. Should be in 2012-1-31 format. 
+#' 
+#' @param folder A string with either a relative or absolute path that contains 
+#'   the trace folders, e.g., 'C:/CRSS/dmi/DNFSinput'.
+#'   
+#' @param start_year The new start_year for the trace files. Should be a single
+#'   numeric value.
+#'   
+#' @param startDate Deprecated. 
 #' 
 #' @export
-crssi_change_nf_start_date <- function(nTrace, folder, startDate)
+crssi_change_nf_start_date <- function(nTrace, folder, start_year, startDate)
 {
-  timeInfo <- paste(startDate,' 24:00', sep = '')
-	
-	# if the new start date is before the old start date, issue a warning
+  assert_that(
+	  length(nTrace) == 1 && is.numeric(nTrace),
+	  msg = "`nTrace` should be a single numeric value."
+	 )
+  
+  assert_that(length(folder) == 1, msg = "`folder` should be a single value.")
+  assert_that(dir.exists(folder), msg = "`folder` does not exist.")
+  
+  assert_that(
+    length(start_year) == 1 && is.numeric(start_year) && start_year > 0 && 
+      start_year <= 9999,
+    msg = "`start_year` should be a single numeric value between 0 and 9999."
+  )
+  
+  assert_that(
+    missing(startDate),
+    msg = "`startDate` is deprecated; please use `start_year` instead."
+  )
+  
+  # if the new start date is before the old start date, issue a warning
 	# but only issue the warning once per call of this function, b/c issuing 
 	# it for every single file and trace would be extremely annoying
 	issueWarning <- FALSE
@@ -67,13 +89,20 @@ crssi_change_nf_start_date <- function(nTrace, folder, startDate)
 			# find the start_date keyword; 
 			startKeyword <- match("start_date", tmpData$atts[1,])
 			
-			#if it doesn't exist, then can just skip this file
-			if(!is.na(startKeyword)) {
+			# if it doesn't exist, then can just skip this file
+			if (!is.na(startKeyword)) {
+			  # parse the current start_date
+			  timeInfo <- simplify2array(strsplit(tmpData$atts[2, startKeyword], "-"))
+			  # replace the year with the new year
+			  timeInfo[1,1] <- as.character(start_year)
+			  # and then recreate the string
+			  timeInfo <- paste(timeInfo, collapse = "-")
+			  
 			  issueWarning <- issueWarning | 
-			    (as.POSIXct(timeInfo) < as.POSIXct(tmpData$atts[2,startKeyword]))
+			    (as.POSIXct(timeInfo) < as.POSIXct(tmpData$atts[2, startKeyword]))
 			  
 			  # replace existing start_date
-			  tmpData$atts[2,startKeyword] <- timeInfo
+			  tmpData$atts[2, startKeyword] <- timeInfo
 			  
 			  # reconstruct the header info
 			  atts <- paste0(tmpData$atts[1,], ": ", tmpData$atts[2,])
@@ -81,7 +110,7 @@ crssi_change_nf_start_date <- function(nTrace, folder, startDate)
 			  # writes out to the same folder it reads in from
 			  utils::write.table(
 			    tmpData, 
-			    file = file.path(currFold,currFiles[j]),
+			    file = file.path(currFold, currFiles[j]),
 			    quote = FALSE, 
 			    row.names = FALSE,
 			    col.names = FALSE
@@ -97,7 +126,54 @@ crssi_change_nf_start_date <- function(nTrace, folder, startDate)
 	    "  Consider using crssi_create_dnf_files() instead."
 	  )
 	
+	update_readme(folder, start_year)
+	
 	invisible(nTrace)
+}
+
+update_readme <- function(folder, start_year)
+{
+  rfile <- file.path(folder, "README.txt")
+  
+  crssio_version <- paste("CRSSIO version:", utils::packageVersion('CRSSIO'))
+  dateCreate <- paste('date modified:', Sys.Date())
+  createBy <- paste('modified by:', Sys.info()[["user"]])
+  startYear <- paste('new start year:', start_year)
+
+  if (file.exists(rfile)) {
+    # README exists, so append to it
+    txt <- rbind(
+      "----------",
+      "The trace files had their start_dates changed by `crssi_change_nf_start_date()`.",
+      dateCreate,
+      createBy,
+      startYear,
+      crssio_version
+    )
+    
+    write(txt, file = rfile, append = TRUE)
+  } else {
+    # README doesn't exist, so create it
+    txt <- rbind(
+      "Natural Flow Data",
+      "These file have an unknown creation source.",
+      "----------",
+      "The trace files had their start_dates changed by `crssi_change_nf_start_date()`.",
+      dateCreate,
+      createBy,
+      startYear,
+      crssio_version
+    )
+    
+    utils::write.table(
+      txt, 
+      file = rfile, 
+      quote = FALSE, 
+      row.names = FALSE, 
+      col.names = FALSE
+    )
+  }
+  invisible(txt)
 }
 
 #' @export
@@ -112,9 +188,9 @@ changeStartDate <- function(nTrace, folder, startDate)
 #' 
 #' Changes the start date in all files in each CRSS_DIR/dmi/evap/trace folder.
 #' 
-#' For evaporation files used in CRSS that are in CRSS_DIR/dmi/Evap, this function
-#' will change their start date, and add 0s on if necessary to extend the evaporation
-#' rates. The files specific to Powell are skipped.
+#' For evaporation files used in CRSS that are in CRSS_DIR/dmi/Evap, this 
+#' function will change their start date, and add 0s on if necessary to extend 
+#' the evaporation rates. The files specific to Powell are skipped.
 #' 
 #' @param startDate New start date. Should be in 2012-1-31 format.
 #' @param nTrace The number of traces to process.